Move helpers on to their own module
